backend/requirements.txt
django==5.2.8
djangorestframework==3.15.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
# psycopg2-binary==2.9.9
requests==2.31.0
sendgrid==6.11.0
posthog==3.6.0
stripe==10.6.0
gunicorn==23.0.0
django-waffle==5.0.0
ShopifyAPI==12.7.0
-e ./libs/ecommerce-integrations-sdk
djangorestframework-api-key==3.1.0
celery==5.3.4
redis==5.0.1
easypost==9.0.0
sentry-sdk==1.40.6
